added another function for final collation and writing clean data
